Initial commit to generalize the importance sampling computations to arbitrary distributions, including bounded distributions.
A few comments:  
1. This has been done in collaboration with Mike. 
The general computation will involve both the importance density 
and the original density in standardized space, but the scope 
has expanded from standard normals.  The importance sampling 
still is generated with respect to normal distributions around 
representative points. 

2.  There are some TO-DOs left: 
- Check all four use cases of importance sampling (local reliability, 
global reliability, PCE/SE, and standalone) for the correct 
bounds and the correct space (X or U) for the data and the model. 
- Check the recentered_density function to make sure the handling 
of design variables is correct. 
- Check the construction of the representative weights.  This still
needs to be fixed from the old use of phi. 

3. NOTE:  we found that the Pecos implementation of phi is strictly 
one-dimensional.  That is, if you give it a beta that has been calculated
over multiple dimensions, the resulting normal pdf is not correct (we 
are not accounting for a 2pi^d scaling, where d is the dimensions).  
We now have a fully general multi-variate PDF with what Mike implemented
in DakotaModel with the continuous_probability_density function. 
Mike and I are reviewing the relevant methods in Dakota to ensure 
we are only calling phi in a uni-variate context. 





git-svn-id: https://software.sandia.gov/svn/dakota/trunk@2328 aa0779c3-6ec2-4b84-a212-a405c134cd8d