Extend logic for stopping iterator and evaluation servers and for creating and deallocating communicator partitions in order to handle the case of idle processors caused by strict adherance to a processors_per_server override. Several cases with idle processors would previously hang and are now operational, although there are still a few cases left to address (mainly, peer iterator scheduling in ConcurrentMetaIterator and SeqHybridMetaIterator).
The basic design is that idle processors enter the normal server mode and are
terminated by the scheduler using a stop_servers() function.  This requires
that they are included in the creation of intra- and inter- communicators but
are just not assigned any work due to their exclusion from the numServers count.


git-svn-id: https://software.sandia.gov/svn/dakota/trunk@2344 aa0779c3-6ec2-4b84-a212-a405c134cd8d