Two changes: 1. The number of representative points around which importance sampling is performed is forced to be one for importanceSamplingType = IS or AIS (import or adapt_import).
2.  The weights on the representative points have been updated to 
reflect the density of those points in U-space. 

The first change causes more of a diff than the second change. 
There are a few tests that are diffing and I will update those 
after discussion with Mike. The bounded case for uniforms 
appears to be working.  I will add this and perhaps another 
bounded distribution to dakota_rosenbrock_importance.in. 



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@2383 aa0779c3-6ec2-4b84-a212-a405c134cd8d