Making an incremental commit on work_directory rather than going back to the drawing board (to reconsider whether allowing a shared work directory for concurrent evals is something we want to support).
Automatically tag directory and/or parameters/results files when concurrent local evaluations are possible.  This is a first cut and will be enhanced and moved to init_communicators to cover the message-passing concurrent case.  Is a bit heavy-handed as really all that's needed is awareness of who should make/remove a directory and that the ultimate params/results are unique...

Also change interfaceSynchronization to an enum with values ASYCHRONOUS_INTERFACE and SYNCHRONOUS_INTERFACE.  The use of _INTERFACE here isn't ideal since the concept applies to a Model, but seemed a decent first pass.



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@2756 aa0779c3-6ec2-4b84-a212-a405c134cd8d