Add exception handling for file removal and rename operations, with options for SILENT, WARN, and ERROR.  (The Boost filesystem operations can throw.  There are more operations that should be protected, but this is a start.)
This should improve behavior in cases Patty observed, where Coliny was catching this exception instead of Dakota and providing unhelpful information.  Specifically on Windows, the analysis_driver process might still have a lock on the parameters or results file when Dakota tries to remove it.

Could consider waiting and trying again on platforms that lock like this.



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@2907 aa0779c3-6ec2-4b84-a212-a405c134cd8d