WIP: Data reader - read sigma data
I also added a test for a bad data file.  This is based on insufficient
file data for the array.  If we allow free formatted data, the opposite
scenario, ie more file data than the size of the array, is permissible.
We can add the ability to enforce exact matches if desired.

git-svn-id: https://software.sandia.gov/svn/dakota/trunk@2991 aa0779c3-6ec2-4b84-a212-a405c134cd8d