Field data reader: add unit test for ExperimentData
This is an intermediate commit to get something in place to iterate with
Brian and Laura.  It appears the first function label gets used as the
filename from which to read the field values.  I don't know if this is
intended ot not.

I don't read sigma values as yet...

git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3029 aa0779c3-6ec2-4b84-a212-a405c134cd8d