Moved the workdir_utils test to the unit_test subdir in preparation for more advanced tests for relative paths to analysis drivers.
Note:  Referring to workdir_utils test as a "unit test" is a bit of a stretch since
it depends on many symbols exposed by libdakota_src, but I think this is a step in
the right direction.  Breaking-out WorkdirHelper, etc. from dakota_src into a
smaller component package would be an improvement, but that requires a dependency
analysis/refactoring task.

ToDo:  Consider adding try/catch blocks to WorkdirHelper::file_op_items(...)



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3036 aa0779c3-6ec2-4b84-a212-a405c134cd8d