Sprint 14 Story: Functional data: Refactor ExperimentData class
add read, accessors and unit testing for field coordinate data

Fields may or may not have coordinate data.  The presence of it is
inferred by whether or not the corresponding file exists.  Also,
a consistency check is made between the number of field data entries
and the number of coordinate data entries.  Coordinate data can have
arbitrary dimension (columns).

git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3038 aa0779c3-6ec2-4b84-a212-a405c134cd8d