Enable support for matrix-valued covariance in nonlinear least squares data transformations.
After discussion with MSE, LPS, JDJ, decided to support all covariance forms by default and throw an error if matrix covariance is present in conjunction with an incomplete residual set (non-uniform active set vector for the residuals).  None, scalar, and diagonal covariances can always be applied, even to an incomplete set of residuals.

Next: enable transformation of gradients and Hessians.

git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3068 aa0779c3-6ec2-4b84-a212-a405c134cd8d