Make piecewise decomposition an option for any of the global surrogates.
Joint work with Mohamed and Ahmad.  Instead of having a surrogate of type VPS, have an option for piecewise_decomposition for any of the global surrogates.  Currently GP, RBF, and polynomial are supported.

For testing, set up a new test file for piecewise surrogates, rather than further growing dakota_uq_users.in.  The one active test recovers the historical behavior for the default polynomial.



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3197 aa0779c3-6ec2-4b84-a212-a405c134cd8d