Generalize POF Darts to work with arbitrary surrogate model pointer.
Joint work with Mohamed and Ahmad.  POF Darts now must point to a model of type surrogate, which itself must have an actual_model_pointer.  POF evaluates the truth model initially, then builds and evaluates the surrogate.  Cleaned up a few controls not needed in the class.

TODO: Leaving DIFFs in dakota_pof 2, 3; the previous default was surrogate order 4, now it is 2, until we add higher order to the surrogate spec.

TODO: Add ability for DataFitSurrModel to cache points previously queried in SURROGATE_BYPASS mode.  Then we won't have to cache in POF and resend them.



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3224 aa0779c3-6ec2-4b84-a212-a405c134cd8d