Sprint 21: Implement PCA for field data responses
Add beginning of ReducedBasis class.  It calls the utility to compute
the SVD for a matrix and stores the resulting left and right singular
vectors and the singular values.

I added a unit test that uses Laura's test matrix to do the SVD and then
reconstructs the original matrix from the stored SVD info.  This provides
an example for the class API along with correctness testing.

git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3452 aa0779c3-6ec2-4b84-a212-a405c134cd8d