Lots of file system changes. We'll see how smart SVN is!
As you create exercises for your module, please place them in numbered folders in exercises/<module name>. Students should be able to copy the contents of the folder to a working area and have everything they need to do your exericse. The exception is the plotting tools, which we'll place somewhere on the PATH.

I will entertain arguments about what do with cantilever. We could either put it on the PATH, too, or put copies in each exercise folder. If the latter, I would encourage you to commit symlinks to it instead of making copies, so that if any bugfixes or minor changes are made, we don't have to worry about updating it everywhere. It might be best to prepend cantilever to the PATH so that it overrides the cantilever that exists in the dakota/test folder, which could head off some confusing bugs in interface scripts.




git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3550 aa0779c3-6ec2-4b84-a212-a405c134cd8d