Initial stab at a CMakeLists.txt file that will sequentially build / install Trilinos (teuchos-only) and Dakota as external projects.  Initial attempts on Linux (RHEL6) and Mac (Mavericks) look promising.
As with other commits related to externalproject builds today, other Dakota
developers should not be impacted by this.  Those that are interested in
trying this new build strategy (and providing feedback) should contact me
for tips on the cmake invocation / Boost settings, etc.

Request review by Brian and/or Dena.



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3629 aa0779c3-6ec2-4b84-a212-a405c134cd8d