Add initial capability to calibrate multipliers on error covariance.  Options are 0, 1, num_experiments, num_responses, or both.  This will calibrate hyperparameters that multiply any user-provided covariance data in the likelihood formulation.
For now, simplified calculations in misfit in a way that looks more like they would in a RecastModel, at the cost of some extra copies, in hopes of moving that direction.  This also gets us closer to supporting presence of data, variance, and multipliers in an independent way, so you can calibrate sigma even when no data, in theory.

Disabling calculation of sample variance from experimental data, since it complicates things, and we don't have a sound strategy for it when field data are present.  If a user wants this, they will have to provide the variance for now.

Outstanding issues TODO:
 * Resolve DIFF in pdakota_cantilever_queso
 * Change from uniform to inverse gamma prior on hyperparameters
 * Update gradient and Hessian transformations to use multiplier hyperparams
 * Calc proper sigma labels from exp * resp for output and update all output fns
 * Make options for data, variance, and hyperparams more independent
Unrelated: support arbitrary priors in DREAM?



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3664 aa0779c3-6ec2-4b84-a212-a405c134cd8d