Explore use of inverse gamma prior on hyperparameters.
This temporary commit is so Laura and I can explore any differences of uniform vs. inverse gamma prior on calibrated error parameters.  To enable, change unifHyperparams(false), and possibly adjust the domain bounds.

It also includes the hyperparameter contributions to the prior or log prior pdf (though these are just a weighting in the uniform case).



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3671 aa0779c3-6ec2-4b84-a212-a405c134cd8d