QUEOS/Bayesian: Cleanup determinant calculations and add gradient/Hessian
Add a convenience function to generate an expanded set of multipliers for use in covariance determinant calculations when hyper-parameters are present.  This makes things a little less efficient, but hopefully a lot cleaner and less error-prine.

Add functions to compute the gradient and Hessian of the log covariance determinant for use in MAP pre-solve.



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3749 aa0779c3-6ec2-4b84-a212-a405c134cd8d