Bug fix: need unique file/directory tags in multistart
In general, when there are concurrent iterators, the same output tagging that is applied to .err and .out files needs to be forwarded to underlying models for use in directory and file tagging to avoid clashes.

The new strategy sets the tags at top-level and recursively on any Iterators, though further Model/Iterator recursions may append to them at run-time.



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3761 aa0779c3-6ec2-4b84-a212-a405c134cd8d