Correct bug in Model::estimate_derivatives() induced by the one-sided finite differencing logic in FDstep1 and FDstep2.  When one-sided finite differences are activated due to interaction with variable bounds, the fn_vals_x0 vector must be defined or else the computed derivatives are erroneous.  The logic for computing/retrieving the initial_map_response was not updated when the one-sided finite differences were added.
I am not certain if this just was a latent bug or if it could induce DIFFs.
(It shows up in dakota_rbdo_short_column when using analytic derivatives at
the design level in combination with MVFOSM + numerical derivatives at the
UQ level, but this combination was not active in the test harness.)
If it does induce any DIFFs, they should be readily blessable as
corrections -- looking into that now.


git-svn-id: https://software.sandia.gov/svn/dakota/trunk@3786 aa0779c3-6ec2-4b84-a212-a405c134cd8d