DAK-758: First pass to get training examples under testing
I have the calibration/cantilever and sen_analysis problems being
tested based on the solution directory with auxiliary files coming from
the exercises directory (except for the calibration data file).

This might not be the cleanest way to approach this, but it's a start.

I generated the baseline results on my Linux box so there may be small
diffs on other platforms.

git-svn-id: https://software.sandia.gov/svn/dakota/trunk@4195 aa0779c3-6ec2-4b84-a212-a405c134cd8d