Initial class to implement Bayesian experimental design using mutual information. The design notes for this are in the collaborative area under AdaptiveBayesExpDesign.docx. This algorithm uses a high-fidelity model to calibrate a low-fidelity model, making use of QUESO and mutual information.  The adaptive design is with respect to the high fidelity model, choosing design points at which the high fidelity model should be run.
The high fidelity model is a function of design variables X, but the low fidelity model is a 
function of both the design variables X and the calibration parameters, theta. 
The algorithm starts with an initial LHS design over X.  Then, we generate a set of 
candidate new potential design to add to the high fidelity model and we choose the candidate 
design which will maximize the information added to the posterior according to the mutual information criteria.  

Brian:  since we talked, I changed this to require two model pointers for the high and low fidelity models. 
I also added a new constructor to the ExperimentData class to initialize the experiment response data. 

Kathryn:  in this commit, I sketched out the code that we discussed to fix the bug with respect to 
prediction interval printing when the variance type is not set.  However, this still needs some work 
to cover all the edge cases so is commented out for now.  



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@4347 aa0779c3-6ec2-4b84-a212-a405c134cd8d